redis added

+++ controllers/notifications.controller.js
@@ -1,3 +1,4 @@
+import redis from "../config/redis.js";
 import Notification from "../models/Notification.js";
 import { decryptText } from "./message.controller.js";
 
@@ -6,6 +7,12 @@ import { decryptText } from "./message.controller.js";
 export const getAllNotifications = async (req, res) => {
     try {
         const { userId } = req.auth();
+
+        const cacheKey = `notifications:${userId}`;
+        const cached = await redis.get(cacheKey);
+        if (cached) {
+            return res.json(JSON.parse(cached));
+        }
 
         await Notification.updateMany({ to_user_id: userId }, {read: true});
         const notifications = await Notification.find({ to_user_id: userId }).populate("from_user_id", "username full_name profile_picture").sort({ createdAt: -1 });
@@ -31,7 +38,10 @@ export const getAllNotifications = async (req, res) => {
             };
         });
 
-        res.json({ success: true, notifications: decryptedNotifications });
+        const responseData = { success: true, notifications: decryptedNotifications }
+        await redis.set(cacheKey, JSON.stringify(responseData), "EX", 30);
+
+        res.json(responseData);
     } catch (error) {
         console.log(error);
         res.json({ success: false, message: error.message });
@@ -42,7 +52,12 @@ export const getAllNotifications = async (req, res) => {
 export const DeleteNotification = async (req, res) => {
     try {
         const {notificationId} = req.params;
+        const notification = await Notification.findById(notificationId);
+        if (!notification) {
+            return res.json({ success: false, message: "Notification not found" });
+        }
         await Notification.findByIdAndDelete(notificationId);
+        await redis.del(`notifications:${notification.to_user_id}`);
         res.json({ success: true, message: "Notification Deleted" });
     } catch (error) {
         console.log(error);
